feat: moved sender to internal
