+++ internal/sender/constants.go
